+++ src/eca/main.clj
@@ -5,9 +5,9 @@
    [babashka.cli :as cli]
    [borkdude.dynaload]
    [clojure.string :as string]
+   [eca.client-http :as client]
    [eca.config :as config]
    [eca.logger :as logger]
-   [eca.client-http :as client]
    [eca.server :as server]))
 
 (set! *warn-on-reflection* true)